+++ packages/app/src/runner/event-manager.ts
@@ -1,15 +1,14 @@
 import Bluebird from 'bluebird'
 import { EventEmitter } from 'events'
-import type { BaseStore } from '@packages/runner-shared/src/store'
+import type { MobxRunnerStore } from '@packages/app/src/store/mobx-runner-store'
 import type { RunState } from '@packages/types/src/driver'
 import type MobX from 'mobx'
 import type { LocalBusEmitsMap, LocalBusEventMap, DriverToLocalBus, SocketToDriverMap } from './event-manager-types'
 import type { AutomationElementId, FileDetails } from '@packages/types'
 
-import { automation } from '@packages/runner-shared/src/automation'
 import { logger } from './logger'
 import type { Socket } from '@packages/socket/lib/browser'
-import { useRunnerUiStore } from '../store'
+import { automation, useRunnerUiStore } from '../store'
 
 export type CypressInCypressMochaEvent = Array<Array<string | Record<string, any>>>
 
@@ -65,7 +64,7 @@ export class EventManager {
     return Cypress
   }
 
-  addGlobalListeners (state: BaseStore, options: AddGlobalListenerOptions) {
+  addGlobalListeners (state: MobxRunnerStore, options: AddGlobalListenerOptions) {
     const rerun = () => {
       if (!this) {
         // if the tests have been reloaded
@@ -134,11 +133,6 @@ export class EventManager {
     })
 
     this.ws.on('specs:changed', ({ specs, testingType }) => {
-      // do not emit the event if e2e runner is not displaying an inline spec list.
-      if (testingType === 'e2e' && state.useInlineSpecList === false) {
-        return
-      }
-
       state.setSpecs(specs)
     })
 
@@ -596,7 +590,7 @@ export class EventManager {
     this.ws.off()
   }
 
-  async teardown (state: BaseStore) {
+  async teardown (state: MobxRunnerStore) {
     if (!Cypress) {
       return
     }
@@ -608,7 +602,6 @@ export class EventManager {
     Cypress.stop()
 
     this.studioRecorder.setInactive()
-    this.selectorPlaygroundModel.setOpen(false)
   }
 
   resetReporter () {
@@ -630,7 +623,7 @@ export class EventManager {
     this.localBus.emit('restart')
   }
 
-  async runSpec (state: BaseStore) {
+  async runSpec (state: MobxRunnerStore) {
     if (!Cypress) {
       return
     }