+++ packages/launchpad/src/setup/InstallDependencies.cy.tsx
@@ -33,8 +33,8 @@ describe('<InstallDependencies />', () => {
   it('shows expected actions', () => {
     cy.contains('button', defaultMessages.clipboard.copy).should('be.visible')
     cy.contains('button', defaultMessages.setupPage.step.back).should('be.visible')
-    cy.tick(180000)
-    cy.contains('button', defaultMessages.setupPage.install.checkForUpdates).should('be.visible')
+
+    cy.contains('button', defaultMessages.setupPage.step.skip).should('be.visible')
   })
 
   it('triggers back button callback', function () {