+++ system-tests/projects/webpack-dev-server/cypress.config.js
@@ -1,0 +1,13 @@
+const { devServer } = require('@cypress/webpack-dev-server')
+
+module.exports = {
+  'component': {
+    supportFile: false,
+    devServer,
+    devServerConfig: {
+      output: {
+        publicPath: '/',
+      },
+    },
+  },
+}